New_backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
motor>=2.5.1
aiohttp>=3.8.1
beautifulsoup4>=4.10.0
python-multipart>=0.0.5
pydantic>=1.9.0
python-dotenv>=0.19.1
scikit-learn>=1.0.2
numpy>=1.22.0 